achievements page tranferred to pages
class new-square rename to completion-banner
width and height - change to pixels
