--- src/app/achievements/achievements.page.html
@@ -1,33 +1,0 @@
-<ion-header>
-  <ion-toolbar>
-    <ion-title>Achievements</ion-title>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <div class="container">
-    <!-- Large square at the top showing completion percentage -->
-    <div class="new-square">
-      <span class="square-text">Complete (-%)</span>
-    </div>
-
-    <!-- Grid of achievements with 20 squares in a 5x4 layout -->
-    <div class="grid-container">
-      <div *ngFor="let square of squares; let index = index"
-           class="square"
-           (click)="toggleText(index)"
-           [class.expanded]="square.expanded">
-        <div class="svg-slot" *ngIf="!square.expanded">
-          <!-- SVG for each square when it is not expanded -->
-          <img [src]="square.svg" alt="SVG Icon" />
-        </div>
-
-        <!-- Expanded view showing title and description -->
-        <div *ngIf="square.expanded">
-          <div class="square-title">{{ square.title }}</div>
-          <div class="square-text">{{ square.text }}</div>
-        </div>
-      </div>
-    </div>
-  </div>
-</ion-content>
